feat(input): 사용자 배팅 금액 입력 및 배팅

+++ src/main/java/domain/Money.java
@@ -17,7 +17,7 @@ public class Money {
         return new Money(amount + value);
     }
 
-    public Money minus(int value) {
+    public Money minus(double value) {
         return new Money(amount - value);
     }
 }
